Reorganize task graph tests
