+++ test/CMakeLists.txt
@@ -12,6 +12,7 @@ set(TEST_TARGETS
   graph_generation_tests
   graph_compaction_tests
   intrusive_graph_tests
+  task_graph_tests
 )
 
 add_library(unit_test_suite unit_test_suite.cc unit_test_suite_celerity.cc)
